+++ chevah/compat/testing/conditionals.py
@@ -100,7 +100,10 @@ def onAdminPrivileges(present):
     tests suite with a regular account.
     """
     hostname = gethostname()
-    is_running_as_normal = ChevahTestCase.os_version in ['nt-5.1', 'nt-5.2']
+    is_running_as_normal = (
+        ChevahTestCase.os_version in ['nt-5.1', 'nt-5.2']
+        or ChevahTestCase.TEST_LANGUAGE == 'FR'
+        )
 
     def check_administrator():
         if present: